+++ src/hooks/reactQuery/comment/useDeleteComment.ts
@@ -1,12 +1,9 @@
-import { usePathname } from 'next/navigation';
-import type { InfiniteData } from '@tanstack/react-query';
 import { useMutation } from '@tanstack/react-query';
 
 import { api } from '@/apis';
-import { useIsMyMap } from '@/hooks';
+import { useUpdateTimelineCommentCount } from '@/features/home/hooks/useUpdateTimelineCommentCount';
 import { useToast } from '@/hooks/useToast';
 
-import type { TimelineResponse } from '../goal/useGetTimeline';
 import { useOptimisticUpdate } from '../useOptimisticUpdate';
 
 import type { CommentResponse } from './useGetComment';
@@ -17,11 +14,7 @@ type CommentDeleteRequest = {
 };
 
 export const useDeleteComment = () => {
-  const pathname = usePathname();
-  const [, , username] = pathname.split('/');
-
-  const { isMyMap } = useIsMyMap();
-  const timelineQueryKey = isMyMap ? ['timeline'] : ['publicTimeline', username];
+  const { queryKey, updateTimelineCommentCount } = useUpdateTimelineCommentCount({ isAddingComment: false });
 
   const { queryClient, optimisticUpdater } = useOptimisticUpdate();
   const toast = useToast();
@@ -41,24 +34,7 @@ export const useDeleteComment = () => {
     onSuccess: (_, { goalId }) => {
       toast.success('댓글을 삭제했어요.');
 
-      queryClient.setQueryData(timelineQueryKey, (old: InfiniteData<TimelineResponse>) => {
-        const newPages = old?.pages.map((page) => ({
-          ...page,
-          contents: page.contents.map((content) =>
-            content.goal.goalId === goalId
-              ? {
-                  ...content,
-                  counts: {
-                    ...content.counts,
-                    comment: content.counts.comment - 1,
-                  },
-                }
-              : content,
-          ),
-        }));
-
-        return { ...old, pages: newPages };
-      });
+      queryClient.setQueryData(queryKey, updateTimelineCommentCount(goalId));
     },
     onError: (_, variable, context) => {
       toast.warning('댓글 삭제에 실패했어요.');
